% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{average}
\alias{average}
\title{Average select columns of matrix together.}
\usage{
average(M, g)
}
\arguments{
\item{M}{numeric matrix}

\item{g}{(character) vector that indicates groups of columns of M}
}
\value{
An N by G matrix M_bar
}
\description{
Given an N by P matrix M, where each column p in 1, ..., P belongs to a group g of G total groups, this function computes the mean column vector per group.
}
\examples{
M <- diag(6)
g <- rep(letters[1:2], 3)
length(g) == ncol(M)
M_bar <- average(M, g)
ncol(M_bar) == length(unique(g))
}
\keyword{computation,}
\keyword{matrix}
