% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaging_matrix.R
\name{averaging_matrix}
\alias{averaging_matrix}
\title{Build averaging matrix from character vector of condition labels.}
\usage{
averaging_matrix(conditions)
}
\arguments{
\item{conditions}{character vector of length n_obs that indicates the condition or class to which each observation
belongs.}
}
\value{
matrix X_bar (n_features by n_classes)
}
\description{
Let X be a data matrix (n_features by n_obs), in which each column j is a pattern vector for observation j, which
belongs to class g.
Applying the averaging matrix A to X computes the class-mean matrix, X_bar (n_features by n_classes), in which each
column is the mean pattern for class g.
}
\details{
NB: this function calls \code{indicator()}, which alphabetically sorts the columns of the output.
}
