% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Mean-center columns of a matrix.}
\usage{
center(x)
}
\arguments{
\item{x}{to-be-centered matrix}
}
\value{
centered matrix
}
\description{
Mean-center columns of a matrix.
}
\examples{
center(row(diag(10)))
}
\keyword{computation,}
\keyword{matrix}
