% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_hemis.R
\name{concat_hemis}
\alias{concat_hemis}
\title{Concatenate gifti data from each hemisphere.}
\usage{
concat_hemis(l, hemis = c("L", "R"), pattern = NULL)
}
\arguments{
\item{l}{a list of giftis. must be of length 2, where each element contains a single gifti, one per hemisphere.}

\item{hemis}{a character vector indicating which list element is L or R. Default order: c("L", "R")}

\item{pattern}{(optional) regexp for extracting only images with matching labels}
}
\value{
matrix of N_vertex by N_images, where N_vertex includes all L and R vertices.
}
\description{
Convenience function for working with objects of class gifti (see gifti package).
}
