% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_matrix.R
\name{contrast_matrix}
\alias{contrast_matrix}
\title{Generates a contrast matrix for forming cross-validated euclidean distance matrix.}
\usage{
contrast_matrix(n, condition_names)
}
\arguments{
\item{n}{The number of conditions among which to estimate pattern contrasts. (Numeric vector of length 1.)}

\item{condition_names}{An optional character vector that gives the name of each condition.
This vector specifies the colnames of the contrast matrix.}
}
\value{
A numeric matrix (n^2 by n).
}
\description{
Primarily for internal use.
}
