% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcor.R
\name{crcor}
\alias{crcor}
\title{Estimate cross-run correlation matrix with downsampling.}
\usage{
crcor(x1, x2, idx1, idx2)
}
\arguments{
\item{x1}{data matrix from fold (run) 1}

\item{x2}{data matrix from fold (run) 1}

\item{idx1}{integer matrix of resampled indices for fold 1 (created with \code{resample_idx()})}

\item{idx2}{integer matrix of resampled indices for fold 2 (created with \code{resample_idx()})}
}
\value{
cross-correlation matrix
}
\description{
Estimate cross-run correlation matrix with downsampling.
}
