% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crcor_rcpp}
\alias{crcor_rcpp}
\title{Estimate cross-run correlation matrix with downsampling.}
\usage{
crcor_rcpp(x1, x2, idx1, idx2, A1, A2)
}
\arguments{
\item{x1}{data matrix}

\item{x2}{data matrix}

\item{idx1}{integer matrix}

\item{idx2}{integer matrix}

\item{A1}{averaging matrix}

\item{A2}{averaging matrix}
}
\description{
For internal use. Call \code{crcor()} instead.
}
