% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvdist.R
\name{cvdist}
\alias{cvdist}
\alias{.cvdist}
\title{Compute the cross-validated Euclidean distance.}
\usage{
cvdist(
  x1,
  x2,
  m = mfutils::contrast_matrix(ncol(x1)),
  nms = NULL,
  center = FALSE,
  scale = FALSE
)

.cvdist(x1, x2, m)
}
\arguments{
\item{x1}{Data matrix (n_features by n_conditions) from fold i}

\item{x2}{Data matrix (n_features by n_conditions) from fold j, j=/=i}

\item{m}{Contrast matrix constructed via \code{contrast_matrix()}.
The default is to compute this matrix on the fly.
However, this arguemnt allows for a contrast matrix to be precomputed and supplied
(i.e., if \code{cvdist} is to be called iteratively on multiple data matrices with the same set of conditions).}

\item{nms}{Optional character vector of condition names.}

\item{center}{Boolean: mean center each condition's pattern prior to computation?}

\item{scale}{Boolean: divisively normalize each condition's pattern to unit length (sqrt(ssq)=1) prior to computation?}
}
\value{
a cross-validated euclidean distance matrix (n_condition by n_condition)
}
\description{
Compute the cross-validated Euclidean distance.
}
