% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcor_r.R
\name{.crcor_r}
\alias{.crcor_r}
\title{Estimate cross-run correlation matrix with downsampling.}
\usage{
.crcor_r(x1, x2, idx1, idx2, A1, A2)
}
\arguments{
\item{x1}{data matrix from fold (run) 1}

\item{x2}{data matrix from fold (run) 1}

\item{idx1}{integer matrix of resampled indices for fold 1 (created with \code{resample_idx()})}

\item{idx2}{integer matrix of resampled indices for fold 2 (created with \code{resample_idx()})}

\item{A1}{averaging matrix for fold 1 (created with \code{averaging_matrix()})}

\item{A2}{averaging matrix for fold 2 (created with \code{averaging_matrix()})}
}
\value{
cross-correlation matrix
}
\description{
Internal function. Call \code{crcor()} instead.
}
