% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist.R
\name{enlist}
\alias{enlist}
\title{Create named list from character vector.}
\usage{
enlist(nms)
}
\arguments{
\item{nms}{(character) names for list}
}
\value{
list
}
\description{
A wrapper for setNames(vector())
}
\examples{
enlist(letters)
}
\keyword{general,}
\keyword{lists}
