% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract data from gifti object.}
\usage{
extract_data(gifti, pattern = NULL)
}
\arguments{
\item{gifti}{object of class gifti from library(gifti)}

\item{pattern}{(optional) regexp for extracting only images with matching labels}
}
\value{
matrix of N_vertex by N_images
}
\description{
Convenience function for working with objects of class gifti (see gifti package).
}
\keyword{fmri,}
\keyword{gifti}
