% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labels.R
\name{extract_labels}
\alias{extract_labels}
\title{Extract labels from a gifti object.}
\usage{
extract_labels(gifti)
}
\arguments{
\item{gifti}{object of class gifti from library(gifti)}
}
\value{
character vector of labels of gifti images
}
\description{
Convenience function for working with objects of class gifti (see gifti package).
}
\keyword{fmri,}
\keyword{gifti}
