% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farout.R
\name{farout}
\alias{farout}
\title{Identify "far outlier" points using IQR method}
\usage{
farout(x, tiles = 4, threshold = 3)
}
\arguments{
\item{x}{numeric vector, matrix, or array}

\item{tiles}{defaults to 4 (the IQR)}

\item{threshold}{defaults to 3 ("far" or "extreme" outliers, i.e., beyond "outer fence")}
}
\value{
boolean vector of \code{length(x)} indicating which points (if any) are outlying
}
\description{
Identify "far outlier" points using IQR method
}
\keyword{computation}
