% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{fold}
\alias{fold}
\title{Average a square matrix over the diagonal}
\usage{
fold(x)
}
\arguments{
\item{x}{numeric square matrix}
}
\value{
numeric square symmetric matrix (\code{xbar})
}
\description{
For a square matrix \code{x}, a 'folded' matrix \code{xbar} will have element
\code{xbar[i, j]} = \code{xbar[j, i]} = (\code{x[i, j]} + \code{x[j, i]})/2
}
\keyword{computation,}
\keyword{matrix}
