% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasser2016.R
\docType{data}
\name{glasser2016}
\alias{glasser2016}
\alias{glasser2016_fsaverage}
\alias{glasser2016_fsaverage5}
\alias{glasser2016_fsaverage6}
\alias{glasser2016_fsLR}
\title{Atlas datasets: glasser2016}
\format{
An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.
}
\usage{
glasser2016_fsaverage

glasser2016_fsaverage5

glasser2016_fsaverage6

glasser2016_fsLR
}
\description{
These datasets contain the Glasser 2016 parcellation atlases, saved as R objects of class "atlas".
Atlas objects are provided for all freesurfer spaces (fsaverage, fsaverage5, fsaverage6), as well as the HCP's fsLR
space (within which the original atlas was defined).
This atlas is only defined in surface spaces.
}
\keyword{datasets}
