% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{indicator}
\alias{indicator}
\title{Create indicator matrix from character vector.}
\usage{
indicator(x)
}
\arguments{
\item{x}{character vector of length n}
}
\value{
Indicator matrix. Column names taken from unique(x).
}
\description{
Rows of an indicator matrix correspond to "observations".
Columns correspond to "conditions" (e.g., levels of a factor).
Each observation "belongs" to only one condition.
This membership is indicated by a 1 in the corresponding column (otherwise 0).
}
\examples{
x <- rep(letters[1:4], each = 4)
M <- indicator(x)
nrow(M) == length(x)
ncol(M) == length(unique(x))
M == stats::model.matrix(~ as.factor(x) + 0)
}
\keyword{computation,}
\keyword{matrix}
