% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_equal.R
\name{is_equal}
\alias{is_equal}
\title{Vectorized test for numeric equivalence.
From Johan Larsson's answer here https://stackoverflow.com/questions/35097815/vectorized-equality-testing}
\usage{
is_equal(x, y, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{tol}{tolerance. defaults to .Machine\$double.eps^0.5}
}
\value{
boolean vector
}
\description{
Vectorized test for numeric equivalence.
From Johan Larsson's answer here https://stackoverflow.com/questions/35097815/vectorized-equality-testing
}
\keyword{computation}
