% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_square.R
\name{is_square}
\alias{is_square}
\alias{is_square_mat}
\alias{is_square_vec}
\alias{is_square_tri}
\title{Check if number, vector, or matrix is 'square'}
\usage{
is_square(x)

is_square_mat(x)

is_square_vec(x)

is_square_tri(x)
}
\arguments{
\item{x}{vector or matrix}
}
\value{
boolean
}
\description{
\code{is_square(x)} checks if a scalar value is a perfect square.
\code{is_square_mat(x)} checks if a matrix \code{x} is square (n by n).
\code{is_square_vec(x)} and \code{is_square_tri(x)} work on vectors.
\code{is_square_vec(x)} tests if the vector could be from/formed into a vectorized square matrix (i.e., all elements),
whereas \code{is_square_tri(x)} tests if the vector could be from a triangle of a square matrix (using quadratic formula).
}
\keyword{matrix}
