% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_atlas.R
\name{load_atlas}
\alias{load_atlas}
\title{Load atlas}
\usage{
load_atlas(name, space)
}
\arguments{
\item{name}{character string of atlas name}

\item{space}{character string of atlas name}
}
\value{
object of type atlas
}
\description{
https://adv-r.hadley.nz/s3.html
}
\examples{
atlas <- load_atlas(name = "schaefer2018_17_200", space = "fsaverage5")
atlas
}
