% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_mat.R
\name{melt_mat}
\alias{melt_mat}
\title{Convert an array or matrix into a long-form \code{data.frame}}
\usage{
melt_mat(x, col = 1, ...)
}
\arguments{
\item{x}{matrix or array}

\item{col}{numeric index or name of column for which the level ordering should be reversed
(defaults to 1; you will likely not need to tweak this)}

\item{...}{additional optional arguments to \code{reshape2::melt()}}
}
\value{
long-form data.frame
}
\description{
Useful for subsequently plotting matrices with ggplot2().
This function is a wrapper for \code{reshape2::melt()}.
\code{reshape2::melt()} is used here because it nicely preserves \code{dimnames(x)} as long-form variable columns.
This wrapper function simply formats the variable columns created by \code{reshape2::melt()}
 to play nice with ggplot2.
 Specifically, it reverses one factor's levels such that the diagonal of the matrix plotted by ggplot2 runs from
 top-left to  bottom-right.
}
