% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Find the mode of a vector.}
\usage{
Mode(x, all.modes = TRUE)
}
\arguments{
\item{x}{a vector}

\item{all.modes}{boolean that indicates whether multiple modes should be returned if they exist}
}
\description{
From Ken Williams' answer: https://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode
}
