% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_atlas.R
\name{new_atlas}
\alias{new_atlas}
\alias{validate_atlas}
\title{Internal functions for constructing and validating objects of type "atlas".}
\usage{
new_atlas(x = list())

validate_atlas(x)
}
\arguments{
\item{x}{a list or object of type atlas}
}
\value{
object of type atlas
}
\description{
Created following suggestions in https://adv-r.hadley.nz/s3.html.
}
