% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offdiags.R
\name{offdiags}
\alias{offdiags}
\title{Extract off-diagonals of square matrix.}
\usage{
offdiags(x)
}
\arguments{
\item{x}{square matrix}
}
\value{
vector of off-diagonal elements
}
\description{
Returns vector of length n^2 - n, where n
}
\keyword{matrix}
