% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcellate.R
\name{parcellate}
\alias{parcellate}
\title{Parcellate an image using an atlas.}
\usage{
parcellate(x, atlas, col_roi = "id")
}
\arguments{
\item{x}{to-be-parcellated data matrix, with vertices/voxels as rows}

\item{atlas}{an atlas object, defined in the same space as the data matrix \code{x}}

\item{col_roi}{character giving the name of the column in \code{atlas$key} that defines the regions of interest
(ROIs) into which the image are parcellated.
See XXX for more information.}
}
\value{
a list of matrices, one per ROI
}
\description{
Parcellate an image using an atlas.
}
\examples{
x <- as.matrix(1:nrow(schaefer2018_7_100_fsaverage5$data))
x_parc <- parcellate(x, schaefer2018_7_100_fsaverage5)
str(x_parc)
}
