% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_melted_mat.R
\name{plot_melted_mat}
\alias{plot_melted_mat}
\title{Quickly plot a melted matrix}
\usage{
plot_melted_mat(
  x,
  .row = "Var1",
  .col = "Var2",
  .value = "value",
  geom = "raster",
  rotate_xaxis_text = TRUE
)
}
\arguments{
\item{x}{a long-form data.frame}

\item{.row}{character, name of column in \code{x} that holds labels for matrix rows}

\item{.col}{character, name of column in \code{x} that holds labels for matrix cols}

\item{.value}{character, name of column in \code{x} that holds numeric values for matrix cells}

\item{geom}{"raster" (for \code{geom_raster}) or "tile" for (\code{geom_tile}), defaults to "raster" for speed, esp.
with larger matrices.}

\item{rotate_xaxis_text}{boolean, should x-axis text (column labels) be rotated 90 degrees? useful with long labels /
large matrices.}
}
\value{
a ggplot object
}
\description{
A convenience function that ggplots numeric matrices with sensible defaults.
This function takes a dataframe \code{x}, possibly the output of \code{melt_mat()}, as input, along with
three strings, \code{.row}, \code{.col}, \code{.value}, which indicate the names of the associated columns
in the data.frame.
}
\details{
#' The default arguments are set to work with the labels that \code{reshape2::melt()} uses when names(dimnames(x)) are not set.
}
