% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.atlas}
\alias{print.atlas}
\title{Print method for "atlas" class}
\usage{
\method{print}{atlas}(x, ...)
}
\arguments{
\item{x}{An object of class "atlas"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for "atlas" class
}
