% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_dim.R
\name{rename_dim}
\alias{rename_dim}
\title{Rename elements of a given dimension within an array}
\usage{
rename_dim(m, new_names, DIM = 2)
}
\arguments{
\item{m}{matrix or array}

\item{new_names}{character}

\item{DIM}{numeric index or name of dimension (if dimensions are named) whose elements will be renamed}
}
\value{
m but with new names
}
\description{
A convenience function for working with matrices or arrays.
}
\examples{
m <- array(1:8, dim = c(3, 3, 2))
rename_dim(m, c("newname1", "newname2"), 3)
}
\keyword{general,}
\keyword{strings}
