% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Dynamically render rmarkdown report with parameters.}
\usage{
render_report(
  name,
  src_dir,
  params = NULL,
  ...,
  envir = new.env(),
  base_dir = here::here("src")
)
}
\arguments{
\item{name}{(character) .rdm file name}

\item{src_dir}{(character) directory within \code{base_dir} that contains \code{name}}

\item{params}{optional list of parameters to pass to \code{rmarkdown::render}}

\item{...}{additional optional parameters for \code{rmarkdown::render}}

\item{envir}{environment in which to run \code{rmarkdown::render} (default: new.env())}

\item{base_dir}{location of \code{src_dir} on machiene, defaults to \code{here::here("src")}}
}
\description{
See here for more information: https://bookdown.org/yihui/rmarkdown/params-knit.html
}
