% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Resample elements of vector.}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{to-be resampled vector}

\item{...}{(optional) additional arguments to sample.int}
}
\value{
a vector with elements resampled
}
\description{
A safer method than sample(x), for cases where length(x) may equal 1 or >1.
See examples within ?sample.int.
}
\examples{
resample(1:20, replace = TRUE)  ## bootstrap
resample(1:20, size = 4)  ## downsample
}
\keyword{computation,}
\keyword{resampling}
