% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_idx.R
\name{resample_idx}
\alias{resample_idx}
\title{Build matrix of resampled indices from character vector of condition labels.}
\usage{
resample_idx(
  conditions,
  n_resamples = 10000,
  resample_to = NULL,
  replace = FALSE
)
}
\arguments{
\item{conditions}{character vector of length n_obs that indicates the condition or class to which each observation
belongs.}

\item{n_resamples}{The number of resamples to draw (defaults to 1,000)}

\item{resample_to}{The resampled 'resolution': the number of observations per condition in the resampled result}

\item{replace}{Draw with replacement (bootstrap)?}
}
\value{
An integer matrix, n_resamples by n_obs
}
\description{
Build matrix of resampled indices from character vector of condition labels.
}
