% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale2unit.R
\name{scale2unit}
\alias{scale2unit}
\title{Divisively normalize a vector or columns of matrix to length 1.}
\usage{
scale2unit(x)
}
\arguments{
\item{x}{to-be-scaled matrix or vector}
}
\value{
scaled matrix or vector
}
\description{
Divides each element of column (vector) by square root of sum of squares.
}
\examples{
scale2unit(row(diag(10)))
}
\keyword{computation,}
\keyword{matrix}
