% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer2018.R
\docType{data}
\name{schaefer2018}
\alias{schaefer2018}
\alias{schaefer2018_7_100_fsaverage}
\alias{schaefer2018_7_200_fsaverage}
\alias{schaefer2018_7_300_fsaverage}
\alias{schaefer2018_7_400_fsaverage}
\alias{schaefer2018_7_500_fsaverage}
\alias{schaefer2018_7_600_fsaverage}
\alias{schaefer2018_7_700_fsaverage}
\alias{schaefer2018_7_800_fsaverage}
\alias{schaefer2018_7_900_fsaverage}
\alias{schaefer2018_7_1000_fsaverage}
\alias{schaefer2018_7_100_fsaverage5}
\alias{schaefer2018_7_200_fsaverage5}
\alias{schaefer2018_7_300_fsaverage5}
\alias{schaefer2018_7_400_fsaverage5}
\alias{schaefer2018_7_500_fsaverage5}
\alias{schaefer2018_7_600_fsaverage5}
\alias{schaefer2018_7_700_fsaverage5}
\alias{schaefer2018_7_800_fsaverage5}
\alias{schaefer2018_7_900_fsaverage5}
\alias{schaefer2018_7_1000_fsaverage5}
\alias{schaefer2018_7_100_fsaverage6}
\alias{schaefer2018_7_200_fsaverage6}
\alias{schaefer2018_7_300_fsaverage6}
\alias{schaefer2018_7_400_fsaverage6}
\alias{schaefer2018_7_500_fsaverage6}
\alias{schaefer2018_7_600_fsaverage6}
\alias{schaefer2018_7_700_fsaverage6}
\alias{schaefer2018_7_800_fsaverage6}
\alias{schaefer2018_7_900_fsaverage6}
\alias{schaefer2018_7_1000_fsaverage6}
\alias{schaefer2018_17_100_fsaverage}
\alias{schaefer2018_17_200_fsaverage}
\alias{schaefer2018_17_300_fsaverage}
\alias{schaefer2018_17_400_fsaverage}
\alias{schaefer2018_17_500_fsaverage}
\alias{schaefer2018_17_600_fsaverage}
\alias{schaefer2018_17_700_fsaverage}
\alias{schaefer2018_17_800_fsaverage}
\alias{schaefer2018_17_900_fsaverage}
\alias{schaefer2018_17_1000_fsaverage}
\alias{schaefer2018_17_100_fsaverage5}
\alias{schaefer2018_17_200_fsaverage5}
\alias{schaefer2018_17_300_fsaverage5}
\alias{schaefer2018_17_400_fsaverage5}
\alias{schaefer2018_17_500_fsaverage5}
\alias{schaefer2018_17_600_fsaverage5}
\alias{schaefer2018_17_700_fsaverage5}
\alias{schaefer2018_17_800_fsaverage5}
\alias{schaefer2018_17_900_fsaverage5}
\alias{schaefer2018_17_1000_fsaverage5}
\alias{schaefer2018_17_100_fsaverage6}
\alias{schaefer2018_17_200_fsaverage6}
\alias{schaefer2018_17_300_fsaverage6}
\alias{schaefer2018_17_400_fsaverage6}
\alias{schaefer2018_17_500_fsaverage6}
\alias{schaefer2018_17_600_fsaverage6}
\alias{schaefer2018_17_700_fsaverage6}
\alias{schaefer2018_17_800_fsaverage6}
\alias{schaefer2018_17_900_fsaverage6}
\alias{schaefer2018_17_1000_fsaverage6}
\title{Atlas datasets: schaefer2018}
\format{
An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.

An object of class \code{atlas} of length 4.
}
\usage{
schaefer2018_7_100_fsaverage

schaefer2018_7_200_fsaverage

schaefer2018_7_300_fsaverage

schaefer2018_7_400_fsaverage

schaefer2018_7_500_fsaverage

schaefer2018_7_600_fsaverage

schaefer2018_7_700_fsaverage

schaefer2018_7_800_fsaverage

schaefer2018_7_900_fsaverage

schaefer2018_7_1000_fsaverage

schaefer2018_7_100_fsaverage5

schaefer2018_7_200_fsaverage5

schaefer2018_7_300_fsaverage5

schaefer2018_7_400_fsaverage5

schaefer2018_7_500_fsaverage5

schaefer2018_7_600_fsaverage5

schaefer2018_7_700_fsaverage5

schaefer2018_7_800_fsaverage5

schaefer2018_7_900_fsaverage5

schaefer2018_7_1000_fsaverage5

schaefer2018_7_100_fsaverage6

schaefer2018_7_200_fsaverage6

schaefer2018_7_300_fsaverage6

schaefer2018_7_400_fsaverage6

schaefer2018_7_500_fsaverage6

schaefer2018_7_600_fsaverage6

schaefer2018_7_700_fsaverage6

schaefer2018_7_800_fsaverage6

schaefer2018_7_900_fsaverage6

schaefer2018_7_1000_fsaverage6

schaefer2018_17_100_fsaverage

schaefer2018_17_200_fsaverage

schaefer2018_17_300_fsaverage

schaefer2018_17_400_fsaverage

schaefer2018_17_500_fsaverage

schaefer2018_17_600_fsaverage

schaefer2018_17_700_fsaverage

schaefer2018_17_800_fsaverage

schaefer2018_17_900_fsaverage

schaefer2018_17_1000_fsaverage

schaefer2018_17_100_fsaverage5

schaefer2018_17_200_fsaverage5

schaefer2018_17_300_fsaverage5

schaefer2018_17_400_fsaverage5

schaefer2018_17_500_fsaverage5

schaefer2018_17_600_fsaverage5

schaefer2018_17_700_fsaverage5

schaefer2018_17_800_fsaverage5

schaefer2018_17_900_fsaverage5

schaefer2018_17_1000_fsaverage5

schaefer2018_17_100_fsaverage6

schaefer2018_17_200_fsaverage6

schaefer2018_17_300_fsaverage6

schaefer2018_17_400_fsaverage6

schaefer2018_17_500_fsaverage6

schaefer2018_17_600_fsaverage6

schaefer2018_17_700_fsaverage6

schaefer2018_17_800_fsaverage6

schaefer2018_17_900_fsaverage6

schaefer2018_17_1000_fsaverage6
}
\description{
These datasets contain the Schaefer 2018 parcellation atlases, saved as R objects of class "atlas".
Atlas objects are provided for all parcellation resolutions (100 parcels, 200, ..., 1000), network solutions (7, 17), \
and freesurfer spaces (fsaverage, fsaverage5, fsaverage6).
Currently, only surface atlases are provided.
}
\keyword{datasets}
