% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squareform.R
\name{squareform}
\alias{squareform}
\title{Vectorize and reshape distance or correlation matrices.}
\usage{
squareform(x, lt = TRUE, corr = FALSE, nms = NULL)
}
\arguments{
\item{x}{square matrix or vector}

\item{lt}{boolean, is only the lower-triangle (TRUE) or full matrix (all values) of interest?
For \code{lt = TRUE} (default): when \code{x} is a vector, only the lower-triangle will be extracted, and when
\code{x} is a vector, it will be assumed that it is a lower-triangle vector.
For convenience, this can be toggled to extract/assume the full matrix (both triangles and diagonal)
by setting \code{lt = FALSE}.}

\item{corr}{boolean (defaults to FALSE, for distance matrices), defines the value of diagonal elements}

\item{nms}{optional (applicable only when vec->mat), character vector that holds names of rows/cols of output matrix}
}
\value{
square matrix or vector
}
\description{
When \code{x} is a distance matrix or correlation matrix (symmetric with diagonal elements equal to 0 or 1,
respectively), \code{squareform()} vectorizes this matrix (i.e., "unwraps" or "flattens") into a vector (i.e., vec->mat).
When \code{x} is a vector (i.e., a distance matrix or correlation matrix) \code{squareform()} reshapes this vector
into the correspoding square matrix (i.e., mat->vec).
}
\keyword{matrix}
