% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Var.R
\name{Var}
\alias{Var}
\title{Efficiently compute row-wise or column-wise variance of a matrix.}
\usage{
Var(x, dim = 1, ...)
}
\arguments{
\item{x}{N by P numeric matrix}

\item{dim}{dimension along which to compute variance (1 == row, 2 == col)}

\item{...}{(optional) arguments for rowSums/colMeans calls (e.g., na.rm)}
}
\value{
a vector containing the variances
}
\description{
Reproduced from David Arenburg's answer here: https://stackoverflow.com/questions/25099825/row-wise-variance-of-a-matrix-in-r
}
\examples{
x <- row(matrix(0, 3, 4))
length(Var(x)) == nrow(x)
length(Var(x, 2)) == ncol(x)
}
\keyword{computation,}
\keyword{matrix}
