% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_idx.R
\name{resample_idx}
\alias{resample_idx}
\title{Generate indices for resampling data.}
\usage{
resample_idx(conditions, n_resamples, resample_to = NULL, replace = FALSE)
}
\arguments{
\item{conditions}{character vector}

\item{n_resamples}{integer}

\item{resample_to}{integer}

\item{replace}{boolean}
}
\value{
integer matrix of resampled indices
}
\description{
Generate indices for resampling data.
}
